+++ examples/simple/books/urls.py
@@ -4,8 +4,6 @@ Urls.
 
 from django.conf.urls import url, include
 
-from rest_framework_extensions.routers import ExtendedDefaultRouter
-
 from .views import (
     AddAuthorsToBookView,
     AuthorListView,
@@ -30,26 +28,43 @@ from .viewsets import (
     PublisherViewSet,
 )
 
+from nine.versions import DJANGO_GTE_2_0
+
+if DJANGO_GTE_2_0:
+    from .rest_framework_extensions_routers_compat import ExtendedDefaultRouter
+else:
+    from rest_framework_extensions.routers import ExtendedDefaultRouter
+
 __all__ = ('urlpatterns',)
 
 
 router = ExtendedDefaultRouter()
-books = router.register(r'books',
-                        BookViewSet,
-                        base_name='book')
-publishers = router.register(r'publishers',
-                             PublisherViewSet,
-                             base_name='publisher')
-profiles = router.register(r'profiles',
-                           ProfileViewSet,
-                           base_name='profile')
-authors = router.register(r'authors',
-                          AuthorViewSet,
-                          base_name='author')
+books = router.register(
+    r'books',
+    BookViewSet,
+    base_name='book'
+)
+publishers = router.register(
+    r'publishers',
+    PublisherViewSet,
+    base_name='publisher'
+)
+profiles = router.register(
+    r'profiles',
+    ProfileViewSet,
+    base_name='profile'
+)
+authors = router.register(
+    r'authors',
+    AuthorViewSet,
+    base_name='author'
+)
 
-proxy_authors = router.register(r'proxy-authors',
-                                AuthorProxyViewSet,
-                                base_name='authorproxy')
+proxy_authors = router.register(
+    r'proxy-authors',
+    AuthorProxyViewSet,
+    base_name='authorproxy'
+)
 
 urlpatterns = [
     url(r'^api/', include(router.urls)),