+++ examples/simple/urls.py
@@ -9,6 +9,8 @@ from django.contrib.staticfiles.urls import staticfiles_urlpatterns
 from django.conf.urls.static import static
 from django.views.generic import TemplateView
 
+from nine import versions
+
 from books import urls as books_urls
 
 __all__ = ('urlpatterns',)
@@ -16,11 +18,19 @@ __all__ = ('urlpatterns',)
 admin.autodiscover()
 
 urlpatterns = []
+urlpatterns_args = []
+
+# Admin URLs
+if versions.DJANGO_GTE_2_0:
+    urlpatterns_args += [
+        url(r'^admin/', admin.site.urls),
+    ]
+else:
+    urlpatterns_args += [
+        url(r'^admin/', include(admin.site.urls)),
+    ]
 
 urlpatterns_args = [
-    # Admin URLs
-    url(r'^admin/', include(admin.site.urls)),
-
     # Books URLs
     url(r'^books/', include(books_urls)),