+++ scripts/prepare_docs.sh
@@ -1,3 +1,4 @@
+#!/usr/bin/env bash
 cat README.rst docs/documentation.rst.distrib > docs/index.rst
 cat ADVANCED_USAGE_EXAMPLES.rst > docs/advanced_usage_examples.rst
 cat CHANGELOG.rst > docs/changelog.rst