+++ scripts/pycodestyle.sh
@@ -1,2 +1,3 @@
+#!/usr/bin/env bash
 reset
-pycodestyle src/rest_framework_tricks/
+pycodestyle src/rest_framework_tricks/ --exclude migrations,south_migrations