+++ scripts/pycodestyle_example.sh
@@ -1,3 +1,3 @@
 #!/usr/bin/env bash
 reset
-pycodestyle examples/simple/ --exclude examples/simple/wsgi.py,examples/simple/books/migrations/,examples/simple/books/tests/,
+pycodestyle examples/simple/ --exclude migrations,south_migrations,examples/simple/books/tests/,